modellicity/pyproject.toml
[tool.poetry]
name = "modellicity"
version = "0.1.0"
description = "Python tools for analyzing data in the credit risk space."
authors = [
    "Basil Singer <basil.singer@modellicity.com>",
    "Ralph Minderhoud <ralph.minderhoud@modellicity.com>",
    "Vincent Russo <vincent.russo@modellicity.com>"
]
maintainers = [
    "Basil Singer <basil.singer@modellicity.com>",
    "Ralph Minderhoud <ralph.minderhoud@modellicity.com>",
    "Vincent Russo <vincent.russo@modellicity.com>",

]
readme = "README.md"
homepage = "https://www.modellicity.com"
repository = "https://gitlab.com/modellicity/modellicity"
keywords = ["modelling", "statistics", "credit", "risk", "analysis"]
packages = [
    { include = "modellicity", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.7.1, <3.10"
matplotlib = "^3.1"
numpy = "^1.21"
pandas = "^1.3"
scikit-learn = "^0.24"
sklearn-json = "*"
scipy = "^1.7"
statsmodels = "*"
poetry = "^0.12.17"
wheel = "^0.36"
tabulate = "^0.8.6"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
flake8-docstrings = "^1.5"
nose = "^1.3"
pep8 = "^1.7"
pylint = "^2.4"
mypy = "*"
coverage = "^4.5"
coveralls = "^1.9"
types-python-dateutil = "^0.1.4"
types-tabulate = "^0.1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
